+++ meda/insert.go
@@ -10,12 +10,12 @@ import (
 const InsertsTableName = "inserts"
 
 type Insert struct {
-	Id               int       `db:"id"`
+	Id               uint64    `db:"id"`
 	Rand             float64   `db:"rand"`
 	Path             string    `db:"path"`
 	ModificationTime time.Time `db:"modification_time"`
-	FileSize         int       `db:"file_size"`
-	LastSeen         int       `db:"last_seen"`
+	FileSize         uint64    `db:"file_size"`
+	LastSeen         uint64    `db:"last_seen"`
 }
 
 func InsertsPrepareInsert(ctx context.Context, preparer NamedPreparerContext) (*sqlx.NamedStmt, error) {
